+++ packages/module-backup/src/server/resourcers/backup-files.ts
@@ -42,7 +42,7 @@ export default {
       const rows = await Promise.all(
         backupFiles.slice((page - 1) * pageSize, page * pageSize).map(async (file) => {
           // if file is lock file, remove lock extension
-          return await Dumper.getFileStatus(file.endsWith('.lock') ? file.replace('.lock', '') : file);
+          return await Dumper.getFileStatus(file.endsWith('.lock') ? file.replace('.lock', '') : file, ctx.app.name);
         }),
       );
 
@@ -105,6 +105,7 @@ export default {
       >ctx.request.body;
 
       const app = ctx.tego as Application;
+      const userId = ctx.state.currentUser?.id;
 
       if (data.method === 'worker' && !app.worker?.available) {
         ctx.throw(500, ctx.t('No worker thread', { ns: 'worker-thread' }));
@@ -123,6 +124,7 @@ export default {
               dataTypes: data.dataTypes,
               appName: ctx.tego.name,
               filename: taskId,
+              userId,
             },
             // 目前限制方法并发为1
             concurrency: 1,
@@ -143,6 +145,7 @@ export default {
             dataTypes: data.dataTypes,
             appName: ctx.tego.name,
             filename: taskId,
+            userId,
           })
           .then((res) => {
             app.noticeManager.notify('backup', { level: 'info', msg: ctx.t('Done', { ns: 'backup' }) });